// packages/dnb-eufemia/src/components/modal/ModalContext.d.ts
/**
 * Web ModalContext Context
 *
 */
import React from 'react';
declare const ModalContext: React.Context<{
    preventClick: any;
    onKeyDownHandler: any;
    id: any;
    title: any;
    hideCloseButton: any;
    closeButtonAttributes: any;
    closeTitle: any;
    setBackgroundColor: any;
    onCloseClickHandler: any;
    contentRef: any;
    scrollRef: any;
    hide: any;
    contentId: any;
    close: any;
}>;
export default ModalContext;